refactor: merge single-method files into their parent handler files
Several packages contained files with a single method (25–48 lines)
that were split from their handler with no logical reason:

  mock/serve_raw_content.go  → mock/handler.go
  mock/serve_file_content.go → mock/handler.go
  proxy/request.go           → proxy/handler.go
  proxy/response.go          → proxy/handler.go (copyResponseData inlined)

Each file is now self-contained; proxy/helpers.go is kept separate as
it holds pure utility functions (copyHeaders, copyCookies) used by the
merged methods.
